project/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.35
pydantic==2.8.2
python-multipart==0.0.9
PyYAML==6.0.2
numpy==1.26.4
pypdf==4.3.1
python-docx==1.1.2
pandas==2.2.2
sentence-transformers==3.0.1
